Cleanup after the SWEVO-2 experiments

+++ src/mase/evorbc/RandomRepertoireGenerator.java
@@ -26,6 +26,8 @@ import mase.evaluation.EvaluationResult;
 import mase.evaluation.MetaEvaluator;
 import mase.stat.PersistentSolution;
 import mase.stat.SolutionPersistence;
+import mase.util.CommandLineUtils;
+import mase.util.FormatUtils;
 import org.apache.commons.compress.archivers.tar.TarArchiveOutputStream;
 import org.encog.neural.networks.BasicNetwork;
 import org.encog.neural.pattern.NeuralNetworkPattern;
@@ -36,6 +38,14 @@ import org.encog.neural.pattern.NeuralNetworkPattern;
  */
 public class RandomRepertoireGenerator {
 
+    public static final String MAX_HIDDEN = "-h";
+    public static final String INPUTS = "-i";
+    public static final String OUTPUTS = "-o";
+    public static final String SIZE = "-size";
+    public static final String JOBS = "-jobs";
+    public static final String OUT = "-out";
+    public static final String FORCE = "-force";
+
     public static void main(String[] args) throws IOException {
 
         ParameterDatabase db = Evolve.loadParameterDatabase(args);
@@ -44,46 +54,61 @@ public class RandomRepertoireGenerator {
         MaseProblem sim = (MaseProblem) db.getInstanceForParameter(prob, null, MaseProblem.class);
         sim.setup(state, prob);
 
-        int maxHidden = 10;
-        int inputs = 7;
-        int outputs = 2;
-        int n = 5000;
-        int jobs = 10;
+        int maxHidden = CommandLineUtils.getIntFromArgs(args, MAX_HIDDEN);
+        int inputs = CommandLineUtils.getIntFromArgs(args, INPUTS);
+        int outputs = CommandLineUtils.getIntFromArgs(args, OUTPUTS);
+        int n = CommandLineUtils.getIntFromArgs(args, SIZE);
+        int jobs = CommandLineUtils.getIntFromArgs(args, JOBS);
+        File out = new File(CommandLineUtils.getValueFromArgs(args, OUT));
+        boolean force = CommandLineUtils.isFlagPresent(args, FORCE);
+        out.mkdirs();
 
         Random rand = new Random();
         for (int j = 0; j < jobs; j++) {
             System.out.println("\n---------- Job " + j + " ----------\n");
-            File f = new File("job." + j + ".finalarchive.tar.gz");
-            TarArchiveOutputStream taos = new TarArchiveOutputStream(new GZIPOutputStream(
-                    new BufferedOutputStream(new FileOutputStream(f))));
+            File f = new File(out, "job." + j + ".finalarchive.tar.gz");
+            if (!f.exists() || force) {
+                TarArchiveOutputStream taos = new TarArchiveOutputStream(new GZIPOutputStream(
+                        new BufferedOutputStream(new FileOutputStream(f))));
 
-            for (int i = 0; i < n; i++) {
-                int h = rand.nextInt(maxHidden + 1);
-                NeuralNetworkPattern pattern = NeuralControllerIndividual.createNetworkPattern(
-                        NeuralControllerIndividual.FEED_FORWARD, inputs, h, outputs, false);
-                BasicNetwork network = (BasicNetwork) pattern.generate();
-                int nParams = network.getStructure().calculateSize();
+                for (int i = 0; i < n; i++) {
+                    int h = rand.nextInt(maxHidden + 1);
+                    NeuralNetworkPattern pattern = NeuralControllerIndividual.createNetworkPattern(
+                            NeuralControllerIndividual.FEED_FORWARD, inputs, h, outputs, false);
+                    BasicNetwork network = (BasicNetwork) pattern.generate();
+                    int nParams = network.getStructure().calculateSize();
 
-                double[] params = new double[nParams];
-                for (int p = 0; p < nParams; p++) {
-                    params[p] = -2 + rand.nextDouble() * 4; // [-2,2]
+                    double[] params = new double[nParams];
+                    for (int p = 0; p < nParams; p++) {
+                        params[p] = -5 + rand.nextDouble() * 10; // weights in [-5,5]
+                    }
+                    network.decodeFromArray(params);
+                    AgentController ac = new NeuralAgentController(network.getFlat());
+                    GroupController gc = new HomogeneousGroupController(ac);
+
+                    EvaluationResult[] eval = sim.evaluateSolution(gc, sim.nextSeed(state, 0));
+
+                    PersistentSolution sol = new PersistentSolution();
+                    sol.setController(gc);
+                    sol.setEvalResults(eval);
+                    sol.setFitness(0);
+                    sol.setOrigin(0, 0, i);
+
+                    SolutionPersistence.writeSolutionToTar(sol, taos);
+                    System.out.print(".");
                 }
-                network.decodeFromArray(params);
-                AgentController ac = new NeuralAgentController(network.getFlat());
-                GroupController gc = new HomogeneousGroupController(ac);
-                                
-                EvaluationResult[] eval = sim.evaluateSolution(gc, sim.nextSeed(state, 0));
-                
-                PersistentSolution sol = new PersistentSolution();
-                sol.setController(gc);
-                sol.setEvalResults(eval);
-                sol.setFitness(0);
-                sol.setOrigin(0, 0, i);
-                
-                SolutionPersistence.writeSolutionToTar(sol, taos);
-                System.out.print(".");
+                taos.close();
+
+                // Generate text
+                File s = new File(out, "job." + j + ".archive.stat");
+                try {
+                    RepertoireToText.toText(f, s);
+                } catch (Exception ex) {
+                    ex.printStackTrace();
+                }
+            } else {
+                System.out.println("Already found. Skipping.");
             }
-            taos.close();
         }
     }
